Re-factored color detection code out of main and into color_detect.  Made writing of the intermediate color detection bmp a macro.
