+++ src/main.c
@@ -9,6 +9,7 @@
 #include "frames.h"
 #include "v4l2uvc.h"
 #include "image_utils.h"
+#include "color_detect.h"
 #include "bitmap.h"
 #include "server.h"
 #include "utils.h"
@@ -25,18 +26,6 @@ typedef enum {
     OUTPUT_FILE_TYPE_BMP
 } output_file_t;
 
-typedef struct {
-    float red;
-    float green;
-    float blue;
-    float norm;
-    float red_norm;
-    float green_norm;
-    float blue_norm;
-    float gr_norm;
-    float gb_norm;
-    float rb_norm;
-} detect_color_t;
 
 const char *p_color_detect_file_name = "detect_color_image.bmp~";
 const char *p_color_detect_file_rename = "detect_color_image.bmp";
@@ -190,114 +179,6 @@ bool parseDetectColor(const char *p_detect_color_string, uint32_t detect_color_l
     printf("Converting %s -> r: %f g: %f b: %f\n", p_detect_color_string, p_detect_color->red, p_detect_color->green, p_detect_color->blue);
 
     return true;
-}
-
-bool calcNorms(detect_color_t* p_detect_color) {
-
-    // avoid divide by zeroes
-    if (p_detect_color->red <= 0.0f) {
-        p_detect_color->red = 0.001f;
-    }
-
-    if (p_detect_color->green <= 0.0f) {
-        p_detect_color->green = 0.001f;
-    }
-
-    if (p_detect_color->blue <= 0.0f) {
-        p_detect_color->blue = 0.001f;
-    }
-
-    // calculate and store detect color parameters
-    p_detect_color->norm = sqrt((p_detect_color->red * p_detect_color->red) + (p_detect_color->green * p_detect_color->green) + (p_detect_color->blue * p_detect_color->blue));
-
-    p_detect_color->red_norm = p_detect_color->red / p_detect_color->norm;
-    p_detect_color->green_norm = p_detect_color->green / p_detect_color->norm;
-    p_detect_color->blue_norm = p_detect_color->blue / p_detect_color->norm;
-
-    p_detect_color->gr_norm = p_detect_color->green_norm / p_detect_color->red_norm;
-    p_detect_color->gb_norm = p_detect_color->green_norm / p_detect_color->blue_norm;
-    p_detect_color->rb_norm = p_detect_color->red_norm / p_detect_color->blue_norm;
-
-    return true;
-}
-
-bool rgb_match(detect_color_t *p_detect_color, uint8_t red, uint8_t green, uint8_t blue, float tolerance) {
-
-    /* Min / max ratio skew */
-    const float norm_min = 1.0f - tolerance;
-    const float norm_max = 1.0f + tolerance;
-
-    detect_color_t in_pix_color;
-
-    in_pix_color.red = red;
-    in_pix_color.blue = blue;
-    in_pix_color.green = green;
-
-    if (!calcNorms(&in_pix_color)) {
-        return false;
-    }
-
-    if (in_pix_color.gr_norm > p_detect_color->gr_norm * norm_min && in_pix_color.gr_norm < p_detect_color->gr_norm * norm_max) {
-        if (in_pix_color.gb_norm > p_detect_color->gb_norm * norm_min && in_pix_color.gb_norm < p_detect_color->gb_norm * norm_max) {
-            if (in_pix_color.rb_norm > p_detect_color->rb_norm * norm_min && in_pix_color.rb_norm < p_detect_color->rb_norm * norm_max) {
-                return true;
-            }
-        }
-    }
-
-    return false;
-}
-
-// assumes pixels packed RGBRGBRGB...3 bytes per pixel
-void rgb_color_detection(uint8_t *p_pix, uint32_t pixSize, int width, int height, detect_color_t *p_detect_color, float detect_tolerance) {
-
-    size_t detect_image_size = width * height;
-
-    if (pixSize != width * height * 3) {
-        printf("%s: invalid pixel buffer size (%u != %d * %d)\n", __func__, pixSize, width, height);
-        return;
-    }
-
-    uint8_t *p_detect_image_start = (uint8_t*)malloc(detect_image_size);
-    uint8_t *p_detect_image = p_detect_image_start;
-
-    // iterate over input image buffer and write 0 if specified color not detected, 1 if detected
-    uint8_t *p_input = p_pix;
-
-    for (uint32_t h=0; h < height; ++h) {
-        for (uint32_t w=0; w < width; ++w) {
-            // decode one pixel
-            uint8_t blue = *p_input++;
-            uint8_t green = *p_input++;
-            uint8_t red = *p_input++;
-
-            if (rgb_match(p_detect_color, red, green, blue, detect_tolerance)) {
-                *p_detect_image++ = 255;
-            } else {
-                *p_detect_image++ = 0;
-            }
-        }
-    }
-
-    FILE* p_file = fopen(p_color_detect_file_name, "w+");
-
-    if (p_file == NULL) {
-        panic("Can't write output image file.");
-        free(p_detect_image_start);
-        return;
-    }
-
-    // write image
-    bmWriteBitmap(p_file, width, height, 1, p_detect_image_start, detect_image_size);
-
-    fflush(p_file);
-    fclose(p_file);
-
-    /* Now that write is complete, rename the file */
-    rename(p_color_detect_file_name, p_color_detect_file_rename);
-
-    free(p_detect_image_start);
-    p_detect_image_start = NULL;
 }
 
 void grab_frame(struct frame_buffer *fb, output_file_t file_type, detect_color_t detect_color, bool b_color_detect, float detect_tolerance) {
